services/knowledge_pool/requirements.txt
pdfplumber>=0.10
psycopg2-binary>=2.9
python-dotenv>=1.0
pyyaml>=6.0
openai>=1.0
